Click through rate prediction - Predict which mobile ad will be clicked
In online advertising, click-through rate (CTR) is a very important metric for evaluating ad performance. As a result, click prediction systems are essential and widely used for sponsored search and real-time bidding. Prediction probability of an ad click will help the company determine the bid that should be made for an ad spot.  Dataset is released by Avazu, Avazu provides a marketplace for digital advertising. The dataset size is relatively large.